Split dispersions in single files each
