+++ src/elli/materials_db.py
@@ -1,5 +1,5 @@
 # Encoding: utf-8
-from .dispersions.dispersions import ConstantRefractiveIndex
+from .dispersions import ConstantRefractiveIndex
 from .materials import IsotropicMaterial
 
 AIR = IsotropicMaterial(ConstantRefractiveIndex(n=1))
